% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKm02a0.R
\name{AKq02a0}
\alias{AKq02a0}
\title{\code{AKq02a0} estimates a0 using the Andreev-Kingkade rule of thumb.}
\usage{
AKq02a0(q0, sex = "m")
}
\arguments{
\item{q0}{a value or vector of values of q0, the death probability for age 0 infants.}

\item{sex}{either "m" or "f"}
}
\value{
a0, the estimated average age at death of those dying in the first year of life, either a single value or a vector of a_0 values.
}
\description{
\code{AKq02a0} is an auxiliary function used by version 6 of the four HMD lifetable functions, \code{ltper_AxN()}, \code{ltcoh_AxN()}, \code{ltperBoth_AxN()}, \code{ltcohBoth_AxN()}.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}
