% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CanadaDeaths2023}
\alias{CanadaDeaths2023}
\title{Number of deaths and population by sex and age groups for Canada in 2023}
\format{
`CanadaDeaths2023`
A data frame with 21 rows and 6 columns:
\describe{
  \item{start}{age in years at start of age interval}
  \item{end}{age in years at end of age interval}
  \item{population.female}{female population in the age interval on 2023-07-01}
  \item{population.male}{male population in the age interval on 2023-07-01}
  \item{deaths.female}{number of females in the age interval dying during 2023}
  \item{deaths.male}{number of males in the age interval dying during 2023}
}
}
\source{
Statistics Canada; deaths \url{https://doi.org/10.25318/1310070901-eng};
population \url{https://doi.org/10.25318/1710000501-eng};
downloaded 2025-01-22.
}
\usage{
CanadaDeaths2023
}
\description{
Number of deaths and population by sex and age groups for Canada in 2023
}
\examples{
CanadaDeaths2023
}
\keyword{datasets}
