% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CanadaMortality}
\alias{CanadaMortality}
\title{Sex-specific mortality rates for Canada in 2023 by age groups}
\format{
`CanadaMortality`
A data frame with 20 rows and 4 columns:
\describe{
  \item{start}{age in years at start of age interval}
  \item{end}{age in years at end of age interval}
  \item{q.female}{female mortality rate, annual probability of
  dying in the age interval}
  \item{q.male}{male mortality rate, annual probability of
  dying in the age interval}
}
}
\source{
\url{https://doi.org/10.25318/1310071001-eng}
downloaded 2025-01-22
}
\usage{
CanadaMortality
}
\description{
Sex-specific mortality rates for Canada in 2023 by age groups
}
\examples{
CanadaMortality
}
\keyword{datasets}
