% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CanadaMortalitySingleYear}
\alias{CanadaMortalitySingleYear}
\title{Sex-specific mortality rates for Canada in 2022--2023 by single years of age}
\format{
`CanadaMortalitySingleYear`
A data frame with 110 rows and 3 columns:
\describe{
  \item{age}{age in years}
  \item{q.female}{female mortality rate, probability of
  dying during the year}
  \item{q.male}{male mortality rate, probability of
  dying during the year}
}
}
\source{
Statistics Canada; \url{https://doi.org/10.25318/1310011401-eng};
downloaded 2025-01-23.
}
\usage{
CanadaMortalitySingleYear
}
\description{
Sex-specific mortality rates for Canada in 2022--2023 by single years of age
}
\examples{
head(CanadaMortalitySingleYear)
tail(CanadaMortalitySingleYear)

}
\keyword{datasets}
