% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/haz2cdf.r
\name{cdf2haz}
\alias{cdf2haz}
\title{Convert a descrete CDF into a descrete hazard function}
\usage{
cdf2haz(x)
}
\arguments{
\item{x}{A numeric vector representing a descrete Cumulative Distribution Function (CDF)}
}
\description{
Convert a descrete CDF into a descrete hazard function.
}

