% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ratesFromIncidenceCounts.r
\name{continuousHazardsFromIncidenceCounts}
\alias{continuousHazardsFromIncidenceCounts}
\title{Compute a continuous hazards matrix}
\usage{
continuousHazardsFromIncidenceCounts(x0, incidence, RR, .handleError = FALSE)
}
\arguments{
\item{x0}{a vector with the distribution of the \strong{observed}
population over (a portrion of) the state space at the start of the life
table step.}

\item{incidence}{The (oberved) incidence of each event (relative to
\code{sum(x0)}).}

\item{RR}{A numeric matrix where each row represents the various
relative risks of arriging at a given node from each
of the other nodes.  Note that loops in the transition
states are not allowed, so most of the elements of \code{RR} will be
zero (including \strong{all} of the diagonal elt's). All RR's
must be positive.}

\item{.handleError}{used internally}

\item{transformation}{the transformation to use in optimizing the rates.
options include 'logOdds', 'logNegativeSquaredInverse',and 'identity',
matched by \code{match.arg}.  (default = 'logOdds').}
}
\description{
Calculate rates on the basis of the entire cohort from rates
expressed as proportions of the *alive* cohort.
}
\note{
When incrementing a life table, it is necessary to know both where the
individuals go, but also where they came from in order to keep track of the
whole population.  \code{continuousHazardsFromIncidenceCounts} solves that problem
by solving for a transiziton (rate) matrix is a helper function for

\code{continuousHazardsTolifeTableStep} is great if you have rates, but what if what you have is
are statistics on the number of events expected in the next interval?

what if you knew the number of BSO's, benign DX's, XXOC's, and cancers,
you epect in the population, and you know the relative risks of these
events within the population, but you don't know the rates.
}
\seealso{
Other lifetable: \code{\link{continuousHazardsToLifeTableStep}};
  \code{\link{countsToLifeTable}};
  \code{\link{distByAge2DiscreteHazards}};
  \code{\link{lifeTable}};
  \code{\link{optimizeSwitchProbs}}
}

