% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ratesToLifeTableStep.r
\name{continuousHazardsToLifeTableStep}
\alias{continuousHazardsToLifeTableStep}
\title{Compute a single increment of a life table.}
\usage{
continuousHazardsToLifeTableStep(x0, M, steps = 1)
}
\arguments{
\item{x0}{A vector with the distribution of the population
across the state space at the start of the life table iteration.}

\item{M}{A matrix with the abosolute (net) risks of transition from any
state in the state-space model to another.}

\item{data}{A data.frame with one row per age interval, ordered by age.}
}
\description{
Compute a single increment (step) of a life table given the initial state
and the transition (Rate) matrix.  As a conveninece, a certian number of
sub-steps (\code{steps}) can be returned.
}
\note{
This implements a solution to a Homogeneous Linear Systemwith Constant
Coefficients described in section 7.5 of "Elementary Differential
Equations and Boundary Value Problems" by Boyce and Diprima, 6th
Edition.

NOTE THAT THIS FUNCTION REQUIRES RATES AND NOT COUNTS OR PROBABILITEIS!
if you have expected counts (i.e. proportions of the \strong{entire}
birth cohort) of event (arrivals at a node) within an
interval, use \code\link{countsToLifeTable}
}
\seealso{
Other lifetable: \code{\link{continuousHazardsFromIncidenceCounts}};
  \code{\link{countsToLifeTable}};
  \code{\link{distByAge2DiscreteHazards}};
  \code{\link{lifeTable}};
  \code{\link{optimizeSwitchProbs}}
}

