% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countsToLifeTable.r
\name{countsToLifeTable}
\alias{countsToLifeTable}
\title{Compute of a life table from count and relative risk data}
\usage{
countsToLifeTable(incidence, x0, RR)
}
\arguments{
\item{incidence}{A data.frame with one row per age interval, ordered by age.}

\item{x0}{A vector with the distribution of the population
across the state space at birth of the cohort.  In the OvaSim model,
states represent the accumulation of events, so typically at birth,
none of the members of the cohort have experienced any events, and the
\code{x0} will be something like:

    \code{c(healty = 1, cancer =0, benign = 0, tubalLidation = 0, dead=0)}

though it doesn't have to be. For example if different RR models
are at different sets of age intervals, then the initial state may
be the final from a previous call to \code{countsToLifeTable}.}

\item{RR}{A numeric matrix where each row represents the various
relative risks of arriging at a given node from each
of the other nodes.  Note that loops in the transition
states are not allowed, so most of the elements of \code{RR} will be
zero (including \strong{all} of the diagonal elt's). All RR's
must be positive.}
}
\description{
Compute (portions) of a life table from count and relative risk data
where counts are expressed as proportions of the \strong{entire} birth
cohort.
}
\note{
In \code{countsToLifeTable()}, 'counts' are really proportions of the
poulation who will experience an event during the age-interval.
(i.e. the fraction of the \strong{entire}
cohort, many of which could be dead before the start of the
current interval.

This is as opposed to most readily available incidence data which are
expressed in terms of "events per women alive at the start of the age
interval", which really describes a rate and not a proportion of the
birth cohort. In the latter case you should use
\code{link{ratesToEventCounts}}
}
\seealso{
Other lifetable: \code{\link{continuousHazardsFromIncidenceCounts}};
  \code{\link{continuousHazardsToLifeTableStep}};
  \code{\link{distByAge2DiscreteHazards}};
  \code{\link{lifeTable}};
  \code{\link{optimizeSwitchProbs}}
}

