% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelParamGenerators.r
\name{decisionRuleWriter}
\alias{decisionRuleWriter}
\title{create a decision rule module for PopSim}
\usage{
decisionRuleWriter(hazardModel, additive = F, file)
}
\arguments{
\item{hazardModel}{a matrix with the hazard model supplied to `lifeTable`}

\item{additive}{indicates whether the subsequent relative hazards are applied additively or if they replace the previously applied HR}

\item{file}{A writable connection or a character string naming the file to write to}
}
\description{
create a decision rule module for PopSim.
}
\examples{
stateNames <- c('H',# healthy
			'TL', # tubes ligated
			'BSO',# tubes and ovaries removed
			'Ben',# diagnosed with Benign Disease
			'Can',# cancer
			'XX') # dead without cancer

hazardModel <- matrix(c(
		0,0    ,0    ,0,0,0,
		1,0    ,0    ,0,0,0,
		1,1-0.2,0    ,0,0,0,
		1,1    ,1    ,0,0,0,
		1,1-0.3,1-0.4,0,0,0,
		1,1    ,1    ,1,0,0),
					  ncol=6,
					  byrow=TRUE,
					  dimnames = list(stateNames,
									  stateNames))

decisionRuleWriter(hazardModel)
}

