% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distByAge2DiscreteHazards.r
\name{distByAge2DiscreteHazards}
\alias{distByAge2DiscreteHazards}
\title{Compute a set of Discrete Hazards}
\usage{
distByAge2DiscreteHazards(lifeTable, stateFields = names(lifeTable),
  ctsHazMxList, sharedParamList, writeFile = file.path(tempdir(),
  paste(format(as.Date(date(), "\%a \%b \%d \%H:\%M:\%S \%Y"), "\%Y\%m\%d"),
  "_distByAge2DiscreteHazards_tmp_outputs.csv", sep = "")))
}
\arguments{
\item{lifeTable}{a data.frame with one row per age, ordered by age, where
each row represents the distribution of the population across the model
states at the \strong{start} of each interval.}

\item{stateFields}{(optional, default: names(lifeTable)) The fields in the
datafraem which contain the counts or probabilies for each state.}

\item{ctsHazMxList}{a list of transition matricies}

\item{sharedParamList}{a set of parameters shared across iterations of
a life table calculateion, created by calling
\code{prepParams(hazNameMX,propHazMx)}}

\item{file}{(optional) Either a character string naming a file or a
connection open for writing.  '""' indicates output to the console.  This
can be useful for monitoring a long running procedure.}
}
\value{
A set of absolute hazards for each state-transition and each age
interval
}
\description{
Compute a set of Discrete Hazards.
}
\details{
This function takes (1) a dataset where each row contains the
distribution of population across the model states at a particular age, and
matricies that specify (2) the transition probabilities that are to be estimated
and (3) the hazard ratios that are to be applied to variaous transitions.
}
\seealso{
Other lifetable: \code{\link{continuousHazardsFromIncidenceCounts}};
  \code{\link{continuousHazardsToLifeTableStep}};
  \code{\link{countsToLifeTable}}; \code{\link{lifeTable}};
  \code{\link{optimizeSwitchProbs}}
}

