% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx2dx.R
\name{dx2e0}
\alias{dx2e0}
\title{dx2e0 calculates e0 given \code{dx}}
\usage{
dx2e0(dx, sex = "m")
}
\arguments{
\item{dx}{vector of single age \code{dx} from the lifetable.}

\item{sex}{\code{"m"} or \code{"f"}. The a0 method is sex-specific.}
}
\value{
e0 single value, not a vector.
}
\description{
Calculates life expectancy at birth given \code{dx}.
}
