% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx2dx.R
\name{dx2lx}
\alias{dx2lx}
\title{dx2lx coverts single-age dx to single age lx.}
\usage{
dx2lx(dx)
}
\arguments{
\item{dx}{vector of single age \code{dx} from the lifetable. This is the density function, and it has various interpretations.}
}
\value{
lx vector of the survival function, of same radix as \code{dx}
}
\description{
This assumes 1) that dx sums to the radix exactly and 2) that the last element of dx is also the last element of lx.
}
