% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dx2mx}
\alias{dx2mx}
\title{dx2mx creates mx vector given \code{dx}}
\usage{
dx2mx(dx, sex = "m")
}
\arguments{
\item{dx}{vector of single age \code{dx} from the lifetable.}

\item{sex}{\code{"m"} or \code{"f"}. The a0 method is sex-specific.}
}
\value{
mx vector of lifetable death rates.
}
\description{
Calculates \code{ax} ad \code{qx} and then uses the identity between \code{ax}, \code{qx} and \code{mx} to calculate \code{mx}.
}

