% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EOC_life_table.r
\docType{data}
\name{EOC_develop}
\alias{EOC_develop}
\title{A life table for Devloping Ovarian Cancer}
\format{A data frame with 20 rows and 7 variables:
\describe{
  \item{AgeIinterval}{Age Interval}
  \item{AliveWithoutCancerAtBegining}{Total Alive And Cancer Free At Beginning Of Interval}
  \item{DevelopCancer}{Number Who Develop Cancer This Interval}
  \item{DeathsWithoutCancer}{Number Who Die From Other Causes This Interval Among cancer-free population at the beginning of the interval who did not develop cancer during the interval. }
  \item{CumProbDevelopCancer}{Cumulative Prob. Of Developing Cancer From Birth}
  \item{CumProbDeathWithoutCancer}{Cumulative Prob. Of Dying Cancer Free From Birth}
  \item{CumProbCancerOrDeathOtherCauses} {Cumulative Prob. Of Developing Cancer Or Dying Cancer Free From Birth}
}}
\source{
{

Software Citation:

DevCan: Probability of Developing or Dying of Cancer Software, Version 6.7.2. Statistical Research and Applications Branch, National Cancer Institute, 2014. \url{http://surveillance.cancer.gov/devcan}

Methods Citations:

Fay, M.P., Pfeiffer, R., Cronin, K.A., Le, C. and Feuer, E.J. (2003)
"Age-conditional probabilities of developing cancer" Statistics in Medicine,
22(11): 1837-1848.

Fay M.P. "Estimating age conditional probability of developing disease from
surveillance data".  Population Health Metrics 2004, 2:6.
\url{http://www.pophealthmetrics.com/content/2/1/6}

Database Citation:

Surveillance, Epidemiology, and End Results (SEER) Program
(www.seer.cancer.gov) DevCan database: "SEER 18 Incidence and Mortality,
2000-2011, with Kaposi Sarcoma and Mesothelioma". National Cancer Institute,
DCCPS, Surveillance Research Program, Cancer Statistics Branch, released
September 2014, based on the November 2013 submission. Underlying mortality
data provided by NCHS (www.cdc.gov/nchs).
}

{

Software Citation:

DevCan: Probability of Developing or Dying of Cancer Software, Version 6.7.2. Statistical Research and Applications Branch, National Cancer Institute, 2014. \url{http://surveillance.cancer.gov/devcan}

Methods Citations:

Fay, M.P., Pfeiffer, R., Cronin, K.A., Le, C. and Feuer, E.J. (2003)
"Age-conditional probabilities of developing cancer" Statistics in Medicine,
22(11): 1837-1848.

Fay M.P. "Estimating age conditional probability of developing disease from
surveillance data".  Population Health Metrics 2004, 2:6.
\url{http://www.pophealthmetrics.com/content/2/1/6}

Database Citation:

Surveillance, Epidemiology, and End Results (SEER) Program
(www.seer.cancer.gov) DevCan database: "SEER 18 Incidence and Mortality,
2000-2011, with Kaposi Sarcoma and Mesothelioma". National Cancer Institute,
DCCPS, Surveillance Research Program, Cancer Statistics Branch, released
September 2014, based on the November 2013 submission. Underlying mortality
data provided by NCHS (www.cdc.gov/nchs).
}
}
\usage{
EOC_develop

EOC_develop
}
\description{
A life table for Devloping Ovarian Cancer in Women of all races generated using SEER data collected 2003-2005 using the DevCan software

A life table for Dying from Ovarian Cancer Vs. other causes in Women of all races generated using SEER data collected 2003-2005 using the DevCan software
}
\keyword{datasets}

