% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/periodEventCountsFromAnnualRates.r
\name{EventCountsFromContinuousHazards}
\alias{EventCountsFromContinuousHazards}
\title{Calculate the number of events (incidence)}
\usage{
EventCountsFromContinuousHazards(x0, M)
}
\arguments{
\item{x0}{A vector with the initial state of the cohort at the
start of the interval}

\item{M}{A Matrix with the absolute, continuous transition paramters}
}
\value{
A vector with the incidence of each event during the interval.
\strong{Note that} in general this is \strong{not} the same as the
differnce between the state of the cohort at the start and end of the
interval.
}
\description{
Calculate the number of events (incidence) expected in an interval
given the intial state and the (absolute, continuous) hazard matrix
}

