% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelParamGenerators.r
\name{generateModelParams}
\alias{generateModelParams}
\title{create XML probabability distributions for PopSim}
\usage{
generateModelParams(ages, netHazards, hazardNames = colNames(netHazards),
  file, baseSeed = format(as.Date(date(), "\%a \%b \%d \%H:\%M:\%S \%Y"),
  "\%Y\%m\%d"))
}
\arguments{
\item{ages}{a vector of ages representing the cut points.  Note that length(ages)
   should equal 1 + nrow(netHazards)}

\item{netHazards}{a matrix composed of columns of net hazards, or an object of class 'LifeTable'}

\item{file}{A writable connection or a character string naming the file to write to.}
}
\description{
create XML probabability distributions for PopSim.
}
\examples{
# using  "life_table_with_Procedures" from vignette("")
generateModelParams(seq(0,100,5),life_table_with_Procedures$netHazards)
generateModelParams(seq(0,100,5),life_table_with_Procedures)
}

