% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EOC_life_table.r
\docType{data}
\name{GyneclogicalSureryRatesPer100k}
\alias{GyneclogicalSureryRatesPer100k}
\title{Rates of Gynecologic Surgeries per 100k}
\format{A data frame with 20 rows and 5 variables:
\describe{
  \item{"ageGroup"}{"ageGroup"}
  \item{"Hysterectomy__abdominal_and_vaginal"}{"Hysterectomy  abdominal and vaginal"}
  \item{"Ligation_of_fallopian_tubes"}{"Ligation of fallopian tubes"}
  \item{"Oophorectomy__unilateral_and_bilateral"}{"Oophorectomy  unilateral and bilateral"}
  \item{"Other_excision_of_cervix_and_uterus"}{"Other excision of cervix and uterus"}
}}
\source{
{

\url{http://hcupnet.ahrq.gov/HCUPnet.jsp?Parms=H4sIAAAAAAAAAF2KMQqAMBAEv5Q7o9EyhYVNEM0H9iSHkWChlb83tg4Mw8I.uHEhTKtpUiFkhZdTnrRKkjL7iMXHERmbT7kmqxINOzFVbbXFb.vHoVrvJMGy6Vzbd.wMj8QvdKc0gW4AAAA3D1C16F1A5B2F2BC7879E9190450A9E5130FAB33&JS=Y}

}
}
\usage{
GyneclogicalSureryRatesPer100k
}
\description{
Rates of Gynecologic Surgeries per 100k (Total Estimated U.S. population, 2003).
}
\note{
{
The unit of analysis for HCUP data is the hospital discharge (i.e., the hospital stay), not a person or patient. more information on Discharges

Rates of discharges are calculated using national population estimates from "Intercensal Estimates of the Resident Population by Single Year of Age, Sex, Race, and Hispanic Origin for the United States: April 1, 2000 to July 1, 2010," Population Division, U.S. Census Bureau, Release date: September 2011. (http://www.census.gov/popest/data/intercensal/national/nat2010.html).
}
}
\keyword{datasets}

