% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/haz2cdf.r
\name{haz2cdf}
\alias{haz2cdf}
\title{Convert a descrete hazard function into a descrete CDF}
\usage{
haz2cdf(x)
}
\arguments{
\item{x}{A numeric vector representing a descrete hazard function}
}
\description{
Convert a descrete hazard function into a descrete CDF.
}

