% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EOC_life_table.r
\docType{data}
\name{HistologyByAge}
\alias{HistologyByAge}
\title{Fractions of EOC subtypes (histologyies) by age}
\format{A data frame with 20 rows and 6 variables:
\describe{
  \item{AgeIinterval}{Age Interval}
  \item{ClearCell}{Proportion of EOC diagnoses with ClearCell histology}
  \item{Endometrioid}{Proportion of EOC diagnoses with Endometrioid histology}
  \item{Misc}{Proportion of EOC diagnoses with Misc histology}
  \item{Mucinous}{Proportion of EOC diagnoses with Mucinous histology}
  \item{Serous}{Proportion of EOC diagnoses with Serous histology}
}}
\source{
{

This is where the reference to SEER goes.

}
}
\usage{
HistologyByAge
}
\description{
Fractions of EOC subtypes (histologyies) by age.
}
\keyword{datasets}

