% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.R
\name{lifetable}
\alias{lifetable}
\title{Construct a life table}
\usage{
lifetable(
  q,
  start = 0:(n - 1),
  end = start,
  radix = 1e+05,
  a0 = 0.7,
  a1 = 0.6,
  annual = TRUE
)
}
\arguments{
\item{q}{age-specific mortality rates}

\item{start}{the age in years at the start of each age interval;
if missing, then the default is single years from \code{0} to 
\code{n - 1}, where \code{end} is the number of intervals}

\item{end}{the age in years at the end of each interval; the
default is \code{start}, which produces single-year intervals;
if the last interval is open-ended then \code{start[n]} can be 
\code{Inf}}

\item{radix}{size of synthetic cohort (default \code{100000})}

\item{a0}{fraction of deaths subtracted in year 0 for computing 
person-years of life \eqn{L_0} (default \code{0.7})}

\item{a1}{fraction of deaths subtracted in year 1 for \eqn{L_1}
(default \code{0.6})}

\item{annual}{if \code{TRUE} (the default) input age-specific 
mortality rates (argument \code{q}) are annual rates even if 
the age intervals aren't 1 year;
output mortality rates \code{q[x]} refer to the entire intervals}
}
\value{
\code{lifeTable()} returns a numeric matrix with rows
labeled by ages (or age intervals) and the following named columns,
where \code{x} is age at the start of each interval:
\code{description}
\item{\code{l[x]}}{number alive at age \code{x}}
\item{\code{d[x]}}{number dying in the interval}
\item{\code{p[x]}}{conditional probability of surviving the interval}
\item{\code{q[x]}}{conditional probability of dying in the interval}
\item{\code{L[x]}}{person-years lived in the interval}
\item{\code{T[x]}}{person-years lived after age \code{x}}
\item{\code{e[x]}}{expectation of life at age \code{x}}
}
\description{
Construct a life table starting with age-specific mortality rates
\eqn{q}. Both single-year and grouped (e.g., 5-year) life tables
are supported. Life tables begin at age 0.
}
\details{
The columns of the life table are computed from \code{q}
under the following assumptions:
\enumerate{
\item Deaths evenly distributed through the year(s) in an interval
(except in years 0 and 1).
\item Those surviving beyond the last interval in the table live 
on average \code{y[n]} more years, where \code{y[n]} is the length of the 
last interval in years; if \code{y[n]} is infinite, then \code{y[n - 1]} is used.
}
}
\examples{
 with(CanadaMortality, 
      lifetable(q.female, start=start, end=end))
 (Lm <- lifetable(CanadaMortalitySingleYear$q.male))
 (Lf <- lifetable(CanadaMortalitySingleYear$q.female))
 plot(0:109, Lm[, "l[x]"]/1000, type="l", col="magenta", 
      lwd=2, lty=2, xlab="Age", ylab="Percent Surviving")
 lines(0:109, Lf[, "l[x]"]/1000, type="l", col="blue",
       lty=1, lwd=2)
 legend("bottomleft", inset=0.02, legend=c("female", "male"),
        lty=1:2, lwd=2, col=c("blue", "magenta"))
 grid()
}
\references{
George Barclay, \emph{Techniques of Population Analysis}, 
Wiley, 1958.
}
