% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.r
\name{logodds}
\alias{logodds}
\title{Transfrom values from the probability to the log-odds scale}
\usage{
logodds(x)
}
\arguments{
\item{x}{a vector of probabilities}
}
\value{
vector of log odds corresponding to \code{x}, the vector of probabilities.
}
\description{
Transfrom values from the probability to the log-odds scale.
}
\examples{
logOddsToProbs(c(.25,.5.75))
}

