% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.r
\name{logOddsToProbs}
\alias{logOddsToProbs}
\title{Transfrom values from the log-odds to the probability scale}
\usage{
logOddsToProbs(x)
}
\arguments{
\item{x}{a vector of log odds}
}
\value{
vector of probabilities corresponding to \code{x}, the vector of log-odds.
}
\description{
Transfrom values from the log-odds to the probability scale.
}
\examples{
logOddsToProbs(c(-1,0,1))
}

