% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx2dx.R
\name{lx2dx}
\alias{lx2dx}
\title{lx2dx coverts single-age lx to single age dx.}
\usage{
lx2dx(lx)
}
\arguments{
\item{lx}{vector of the single-age survival function values}
}
\value{
dx vector of single age \code{dx} from the lifetable. This is the density function, and it has various interpretations. Sums to radix.
}
\description{
\code{dx} will sum to the \code{lx} radix. Assumes but does not check for monontonicity in \code{lx}, so negative values can be produced, but are actually impossible.
}
