% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lx2qx}
\alias{lx2qx}
\title{lx2qx coverts single-age lx to single age qx.}
\usage{
lx2qx(lx)
}
\arguments{
\item{lx}{vector of the survival function from age 0 to omega.}
}
\value{
qx vector of single age conditional death probabilities
}
\description{
This assumes that lx contains all ages, and it also closes the \code{qx} vector out with a value of 1, which is common practice. \code{lx} may be of any radix. No checking is done for monotanicity or for 0s in the denominator. Please be aware that 0s in \code{lx}, except for in the last age, will mess this up, and smooth your data if necessary.
}

