% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimizeSwitchProbs.r
\name{optimizeSwitchProbs}
\alias{optimizeSwitchProbs}
\title{optimizes switch probabilities (hazards)}
\usage{
optimizeSwitchProbs(startState, endState, paramGuess, hazardsMX,
  sharedParamList)
}
\arguments{
\item{startState}{The distribution of individuals in each model state
prior to the transition interval}

\item{endState}{The distribution of individuals in each model state
after to the transition interval}

\item{sharedParamList}{(optional) a set of parameters shared across iterations of
a life table calculateion, created by calling
\code{prepParams(hazNameMX,propHazMx)}}

\item{start}{(optional) An initial guess at the transition probabilities}

\item{hazNameMX}{}

\item{propHazMx}{}

\item{verbose}{If TRUE, runtime updates are printed to the terminal
and get_pAB_matrix(...,verbose=TRUE) is called on the first iteration
of the Newton-Raphson root-finding method (default= FALSE).}
}
\description{
optimizes switch probabilities for a single iteration of a life
table.
}
\seealso{
Other lifetable: \code{\link{continuousHazardsFromIncidenceCounts}};
  \code{\link{continuousHazardsToLifeTableStep}};
  \code{\link{countsToLifeTable}};
  \code{\link{distByAge2DiscreteHazards}};
  \code{\link{lifeTable}}
}

