% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepParams.r
\name{prepParams}
\alias{prepParams}
\title{generate paramters for a set of life table calculaions}
\usage{
prepParams(hazNameMX, propHazMx)
}
\arguments{
\item{hazNameMX}{A parameter matrix (character) see example below for details}

\item{propHazMx}{A proportional hazards matrix}
}
\description{
generate paramters for a set of life table calculaions.
}

