% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qx2dx}
\alias{qx2dx}
\title{qx2dx coverts single-age qx to single age dx.}
\usage{
qx2dx(qx)
}
\arguments{
\item{qx}{vector of single age conditional death probabilities}
}
\value{
dx vector of single age \code{dx} from the lifetable.
}
\description{
The resulting \code{dx} vector will sum to 1 (correspond to a radix-1 lifetable)
}

