% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lx2dx.R
\name{qx2lx}
\alias{qx2lx}
\title{lqx2lx coverts single-age qx to single age lx.}
\usage{
qx2lx(qx)
}
\arguments{
\item{qx}{vector of single age conditional death probabilities}
}
\value{
lx vector of the radix-1 survival function from age 0 to omega.
}
\description{
This assumes that \code{qx} starts at age 0. \code{lx} is created with a radix of 1, so rescale outside of this function if needed. No \code{NA}s or \code{Inf}s!
}
