\docType{data}
\name{UKRmales1965}
\alias{UKRmales1965}
\title{Example data for Ukraine males, 1965, extracted from the Human Mortality Database.}
\source{
  Human Mortality Database, as downloaded in May, 2010.
  \url{www.mortality.org}.

  Human Mortality Database, as downloaded in May, 2010.
  \url{www.mortality.org}.

  Human Mortality Database, as downloaded in May, 2010.
  \url{www.mortality.org}.

  Human Mortality Database, as downloaded in May, 2010.
  \url{www.mortality.org}.

  Human Mortality Database, as downloaded in May, 2010.
  \url{www.mortality.org}.
}
\description{
  single-age data from the HMD for Deaths (count)
  Population (exposure) and central death rates,
  death/exposure.

  single-age data from the HMD for Deaths (count)
  Population (exposure) and central death rates,
  death/exposure.

  single-age data from the HMD for Deaths (count)
  Population (exposure) and central death rates,
  death/exposure.

  single-age data from the HMD for Deaths (count)
  Population (exposure) and central death rates,
  death/exposure.

  single-age data from the HMD for Deaths (count)
  Population (exposure) and central death rates,
  death/exposure.
}
\details{
  All four columns are numeric. Mx provided is not exactly
  equal to D(x)/N(x) due to rounding.

  All four columns are numeric. Mx provided is not exactly
  equal to D(x)/N(x) due to rounding.

  All four columns are numeric. Mx provided is not exactly
  equal to D(x)/N(x) due to rounding.

  All four columns are numeric. Mx provided is not exactly
  equal to D(x)/N(x) due to rounding.

  All four columns are numeric. Mx provided is not exactly
  equal to D(x)/N(x) due to rounding.
}
\references{
  Human Mortality Database. University of California,
  Berkeley (USA), and Max Planck Institute for Demographic
  Research (Germany). Available at \url{www.mortality.org}
  or \url{www.humanmortality.de} (data downloaded on 10
  May, 2010).

  Human Mortality Database. University of California,
  Berkeley (USA), and Max Planck Institute for Demographic
  Research (Germany). Available at \url{www.mortality.org}
  or \url{www.humanmortality.de} (data downloaded on 10
  May, 2010).

  Human Mortality Database. University of California,
  Berkeley (USA), and Max Planck Institute for Demographic
  Research (Germany). Available at \url{www.mortality.org}
  or \url{www.humanmortality.de} (data downloaded on 10
  May, 2010).

  Human Mortality Database. University of California,
  Berkeley (USA), and Max Planck Institute for Demographic
  Research (Germany). Available at \url{www.mortality.org}
  or \url{www.humanmortality.de} (data downloaded on 10
  May, 2010).

  Human Mortality Database. University of California,
  Berkeley (USA), and Max Planck Institute for Demographic
  Research (Germany). Available at \url{www.mortality.org}
  or \url{www.humanmortality.de} (data downloaded on 10
  May, 2010).
}
\keyword{datasets}

