% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EOC_life_table.r
\docType{data}
\name{US_population_2003}
\alias{US_population_2003}
\title{United States National Intercensal Estimates (2003) by Age and Sex}
\format{A data frame with 20 rows and 3 variables:
\describe{
  \item{all}{Intercensal Estimates (as of July 1,2003)}
  \item{male}{Intercensal Estimates of Male (as of July 1,2003)}
  \item{female}{Intercensal Estimates (as of July 1,2003)}
}}
\source{
{

\url{http://www.census.gov/popest/data/intercensal/national/nat2010.html?JS=Y}

}
}
\usage{
US_population_2003
}
\description{
United States National Intercensal Estimates (2003) by Age and Sex.
}
\keyword{datasets}

