% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LifeTable.r
\name{lifeTable}
\alias{lifeTable}
\title{Compute life table}
\usage{
lifeTable(initialState, hazardModel = hazard, hazard, netHazards,
  hType = "continuous", incidence, at.risk, finalState, steps = 1,
  checkNames = FALSE, quietly = FALSE)
}
\arguments{
\item{initialState}{A vector with the distribution of the population
across the state space at the start of the life table, or a life table
that is to be extended.}

\item{incidence}{If provided, the this specifies the incidences [number
of arrivals at each node] during the interval.  The incidence may be
expressed as a proportion of the cohort that is expected to undergoe the event
during the interval (default) or as a fraction of those at risk for the event.
To use the latter, the paramter 'at.risk' must be specified.

Note that if \code{incidence} is a matrix or data.frame, rows in \code{incidence}
are considered incidence at consecutive intervals in the life table.}

\item{at.risk}{Porportion of the birth cohort \code{sum(initialState)} that
was used as the denominator for each each element of the parameter
\code{incidence}. This can be specfied by a nueric value (default =
\code{sum(initialState)}), a vector with the proportion of the cohort used for
each individual parameter, or a function with takes a single parameter
with the current distribution of the cohort (begining with statstate), and returns
a vector withthe proporitn of the cohort that represents the denominator
for each incidence parameter in the current iteration of the life table.}

\item{steps}{The number of records to add to the life table, for each row of
'incidence' or each value of hasMX. If \code{incidence > 1}, then
\code{incidence > 1} intermediate records are added to the returned life table}

\item{checkNames}{If \code{TRUE}, names / dimnames of the various inputs are checked
for consistency. (default=\code{FALSE})}

\item{data}{A data.frame with one row per age interval, ordered by age.}

\item{hazMX}{If the parameter \code{incidence} is sppecified, \code{hazMX} is
treated as a matrix with with the \strong{relative} hazards of transitioning
between states, otherwise \code{hazMX} is treated as the  abosolute (net)
hazards of transitioning between states.}
}
\description{
Compute or extend a life table from Hazard Ratios and incidence, or
absolute (net) hazards.
}
\note{
This implements a solution to a Homogeneous Linear Systemwith Constant
Coefficients described in section 7.5 of "Elementary Differential
Equations and Boundary Value Problems" by Boyce and Diprima, 6th
Edition in order to solve the realted continuous time markov model.
}
\seealso{
Other lifetable: \code{\link{continuousHazardsFromIncidenceCounts}};
  \code{\link{continuousHazardsToLifeTableStep}};
  \code{\link{countsToLifeTable}};
  \code{\link{distByAge2DiscreteHazards}};
  \code{\link{optimizeSwitchProbs}}
}

